reorganize source code
